Remove versions on VAPSets.  Remove HasVersion.
The logic behind having a version on VAPsets was to allow a user to update their *canonical*
beliefs (in the form of sets of value and predictions).  And to surface these updated beliefs
as newer versions via the UI.

However the following problem arises.

You have a VAPset with multiple versions,  The versions were made on the 1st, 12th of the month.
You realise that:
    1.  actually there was also another version you had used on the 5th
    2.  that is currently absent
    3.  you want to record this version now, perhaps because it is important for a discussion
        debugging a decision

With the current version system you would end up with:
    * created_at: 1st, version 1
    * created_at: 12th, version 2
    * created_at: 15th, custom_created_at: 5th, version 3
But the version number is redudant information as it is duplicating (in a less useful way) the created_at info.

Con for removing versions: 2 VAPsets could be made at exactly the same time and a version could disambiguate them.
1. This is improbable
2. The current system of versions would not help (as presumably both processes would check at the same time,
find the version of N, then make new entries with the same created_at, and version of N + 1).

Another pro for removing versions:
They are not actually used in the UI to present the elements as you first filter by and then order by
custom_created_at || created_at.  You do not need versions.