Simplify by removing WComponentStateValue's owner_wcomponent_id
Have removed this "owner_wcomponent_id" attribute as this was part of an
exploratory prototype and had not yet proved its utility.
I think the function this was trying to fill was to associate any particular
StateValue with some higher level component and thus allow multiple StateValues
to be easily pulled into a calculation or simulation by using a SQL query like:

SELECT * FROM wcomponents WHERE json->>'owner_wcomponent_id' = "uuid-of-wcomponent-of-knowledge-view";

However I think we will instead accomplish this by going from a knowledge view
and then pulling in all of its components and those of its base knowledge views
to build a composed knowledge view that will therefore contain all of the
StateValues in it as well as the components they target and overwrite the values of.

An example of where this initial approach using the "owner_wcomponent_id" was
started to be explored is with the "UK Average Home HVAC" e.g. with the
StateValue component that overrides the "Name" component and was (still has) the
"owner_wcomponent_id" set to the wcomponent associated with the knowledge view
that it is contained in:
https://datacurator.org/app/#wcomponents/47d4f426-d78e-40a2-a9f9-6c3fdb78d2d8/&storage_location=27&subview_id=a30b3a3f-bd0a-4007-a182-2d03de8f06fb&view=knowledge&x=948&y=-1767&zoom=44&sdate=2023-08-15&stime=19:05:41&cdate=2023-08-15&ctime=19:13:41